new location for test_pe_win_x8664_driver
